+++ src/Widgets/Text.php
@@ -16,17 +16,22 @@ namespace Konekt\AppShell\Widgets;
 
 use Konekt\AppShell\Contracts\Theme;
 use Konekt\AppShell\Contracts\Widget;
+use Konekt\AppShell\Theme\ThemeColor;
 use Konekt\AppShell\Traits\RendersThemedWidget;
 use Konekt\AppShell\Traits\ResolvesSubstitutions;
+use Konekt\AppShell\Widgets\Concerns\CalculatesContextualColors;
 use Konekt\AppShell\Widgets\Concerns\HasModifier;
 use Konekt\AppShell\Widgets\Concerns\SupportsConditionalRendering;
 
 class Text implements Widget
 {
+    use CalculatesContextualColors;
     use RendersThemedWidget;
     use ResolvesSubstitutions;
     use SupportsConditionalRendering;
     use HasModifier;
+
+    protected const DEFAULT_WRAP = 'span';
 
     protected static $allowedTagAttributes = [
         'class',
@@ -47,6 +52,12 @@ class Text implements Widget
 
     private string $suffix = '';
 
+    private null|array|string $colorDef = null;
+
+    private ?string $hexColor = null;
+
+    private ?string $themeColor = null;
+
     public function __construct(Theme $theme, callable $text, ?string $wrap = null)
     {
         $this->theme = $theme;
@@ -66,7 +77,7 @@ class Text implements Widget
         if (isset($options['bold'])) {
             $instance->bold = (bool) $options['bold'];
             if ($instance->bold && null === $instance->wrap) {
-                $instance->wrap = 'span';
+                $instance->wrap = self::DEFAULT_WRAP;
             }
         }
 
@@ -77,6 +88,8 @@ class Text implements Widget
         if (isset($options['suffix'])) {
             $instance->suffix = (string) $options['suffix'];
         }
+
+        $instance->colorDef = $options['color'] ?? null;
 
         foreach (self::$allowedTagAttributes as $allowedTagAttribute) {
             if (isset($options[$allowedTagAttribute])) {
@@ -97,10 +110,26 @@ class Text implements Widget
 
         $text = $this->text;
 
+        $color = $this->parseColorDefinition(
+            is_string($this->colorDef) ? $this->resolveSubstitutions($this->colorDef, $data) : $this->colorDef,
+            $data,
+            ThemeColor::NONE,
+            true,
+        );
+
+        if (null !== $color->style) {
+            $this->tagAttributes['style'] = ($this->tagAttributes['style'] ?? '') . $color->style;
+        }
+
+        if ($color->isNotEmpty() && null === $this->wrap) {
+            $this->wrap = self::DEFAULT_WRAP;
+        }
+
         return $this->renderViewFromTheme('text', [
             'text' => $this->modify($text($data, $this)),
             'wrap' => $this->wrap,
             'tagAttributes' => $this->tagAttributes,
+            'color' => $color,
             'bold' => $this->bold,
             'prefix' => $this->resolveSubstitutions($this->prefix, $data),
             'suffix' => $this->resolveSubstitutions($this->suffix, $data),