+++ src/resources/views/widgets/text.blade.php
@@ -2,5 +2,8 @@
 if ($bold) {
     $tagAttributes['class'] = ($tagAttributes['class'] ?? '') . 'font-weight-bold';
 }
+if (!$color->themeColor->isNone() ) {
+    $tagAttributes['class'] = ($tagAttributes['class'] ?? '') . 'text-' . $color->themeColor->value();
+}
 ?>
 @if($wrap ?? null)<{{$wrap}}@foreach($tagAttributes as $attrKey => $attrVal) {{ $attrKey }}="{{ $attrVal }}"@endforeach>@endif{{ $prefix }}{{ $text }}{{ $suffix }}@if($wrap ?? null)</{{$wrap}}>@endif