+++ Changelog.md
@@ -5,6 +5,7 @@
 
 - Added Components based on the Laravel Blade Components Feature
 - Added `page-actions` blade stack that themes need to define
+- Added colored enum icon feature via the `color_enum_icon()` helper 
 - Renamed the stylesheet of the default theme from `appshell` to `appshell3`
 - Changed the AppShell Theme - it has been rewritten from scratch with new visual style
 - Changed Breadcrumbs dependency from v7 to v8